requirements.txt
pyjwt==2.10.1
requests==2.32.5
python-telegram-bot==22.4
regex==2025.9.1
python-dotenv==1.1.1
pycryptodome==3.23.0
sentence-transformers==5.1.0
ftfy==6.3.1
python-Levenshtein==0.27.1
langchain-community==0.3.29
langchain==0.3.27
boto3==1.40.32
yandex_cloud_ml_sdk==0.15.0
faiss-cpu==1.12.0
textattack==0.3.8
transformers==4.46.1
torch==2.5.1
